Adds test for svg text, updates deps
As we can expect selection to work on anything that has a selectable
text as part of the DOM (sorry canvas), svg could be handled as well
as demonstrated by the tests. This commit makes this explicit.

Peer dependencies are explicitly added to `package.json` to not rely
on global dependencies.